requirements.txt
streamlit>=1.29.0
openai>=1.0.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

